+++ prost-reflect/src/dynamic/mod.rs
@@ -1,4 +1,5 @@
 mod fields;
+mod fmt;
 mod message;
 #[cfg(feature = "serde")]
 mod serde;
@@ -362,6 +363,26 @@ impl DynamicMessage {
     {
         let buf = self.encode_to_vec();
         T::decode(buf.as_slice())
+    }
+
+    /// Formats this dynamic message using the protobuf text format.
+    ///
+    /// Output is pretty-printed with each field on a new line, and nested messages indented. This function is equivalent to
+    /// formatting the message using the alternate format specifier: `format!("{:#}", message)`.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use prost::Message;
+    /// # use prost_types::FileDescriptorSet;
+    /// # use prost_reflect::{DynamicMessage, DescriptorPool, Value};
+    /// # let pool = DescriptorPool::decode(include_bytes!("../file_descriptor_set.bin").as_ref()).unwrap();
+    /// # let message_descriptor = pool.get_message_by_name("package.MyMessage").unwrap();
+    /// let dynamic_message = DynamicMessage::decode(message_descriptor, b"\x08\x96\x01\x1a\x02\x10\x42".as_ref()).unwrap();
+    /// assert_eq!(dynamic_message.to_string_pretty(), "foo: 150\nnested {\n  bar: 66\n}");
+    /// ```
+    pub fn to_string_pretty(&self) -> String {
+        format!("{:#}", self)
     }
 }